+++ packages/androidx_graphics_shapes/lib/src/material_shapes.dart
@@ -3,5 +3,6 @@ export 'material_shapes/material_shapes.dart'
         RoundedPolygonExtension,
         MorphExtension,
         MaterialShapes,
+        PathBorder,
         RoundedPolygonBorder,
         MorphBorder;